🐛 fix: patchDeviceToken 이동 #269
